Moved definition of context to own file
